refactor(cdk:forms): split the controls.ts (#1198)
