+++ packages/cdk/forms/__tests__/formControl.spec.ts
@@ -1,7 +1,7 @@
 import { flushPromises } from '@vue/test-utils'
 
-import { FormControl } from '../src/controls'
 import { zhCNMessages } from '../src/messages/zh-CN'
+import { FormControl } from '../src/models/formControl'
 import { ValidateErrors } from '../src/types'
 import { Validators } from '../src/validators'
