+++ packages/pro/form/docs/Index.zh.md
@@ -6,3 +6,4 @@ title: ProForm
 subtitle: 高级表单
 ---
 
+此组件功能尚未稳定，请谨慎使用。
