+++ tests/setup.ts
@@ -6,3 +6,6 @@ import * as allIcon from '@idux/components/icon/src/definitions'
 addIconDefinitions(Object.values(allIcon))
 
 global.ResizeObserver = ResizeObserver
+
+const { getComputedStyle } = window
+window.getComputedStyle = elt => getComputedStyle(elt)
