+++ tsconfig.json
@@ -4,16 +4,15 @@
     "outDir": "dist",
     "declarationDir": "dist",
     "lib": ["ESNext", "DOM"],
-    "target": "ES2020",
+    "target": "ESNext",
     "moduleResolution": "node",
     "jsx": "preserve",
-    "jsxFactory": "h",
-    "jsxFragmentFactory": "Fragment",
     "sourceMap": true,
     "declaration": true,
     "downlevelIteration": true,
-    "importHelpers": true,
     "esModuleInterop": true,
+    "importHelpers": true,
+    "isolatedModules": true,
     "resolveJsonModule": true,
     "removeComments": false,
     "strict": true,
@@ -33,5 +32,5 @@
     },
     "types": ["vitest/globals"]
   },
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules", "coverage", "dist"]
 }
